补全权限审核逻辑及job缓存机制

+++ bakss-buz/src/main/java/com/bakss/server/domain/BakssBackupValidate.java
@@ -32,7 +32,7 @@ public class BakssBackupValidate extends BaseEntity
 
     /** $column.columnComment */
     @Excel(name = "${comment}", readConverterExp = "$column.readConverterExp()")
-    private Long expType;
+    private Integer expType;
 
     /** $column.columnComment */
     @Excel(name = "${comment}", readConverterExp = "$column.readConverterExp()")
